+++ ChangeLog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-## [1.10.34] - 2023-04-xx
+## [1.10.34] - 2023-04-11
 
 ### Added